Renamed YuvVideo class to Yuv420Video class
