+++ CMakeLists.txt
@@ -4,11 +4,11 @@ PROJECT(TestProject_Elecard)
 SET(HEADERS bitmapTypes.h
             BmpImage.h
             Yuv420Image.h
-            YuvVideo.h)
+            Yuv420Video.h)
 
 SET(SOURCES BmpImage.cpp
             Yuv420Image.cpp
-            YuvVideo.cpp)
+            Yuv420Video.cpp)
 
 ADD_EXECUTABLE(${PROJECT_NAME} main.cpp ${SOURCES} ${HEADERS})
 TARGET_INCLUDE_DIRECTORIES(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
